Find out which other-language wikis have linked pages

All pages passed to the function must be from the same language edition of Wikipedia. For example, if you have a set of English pages, you can find all the interwiki links for those pages. If you have a set of pages from disparate language editions, you can segment the data using dplyr::group_by or similar.

Usage

get_langlinks(titles, language = "en")

Arguments

titles

A character vector of page titles

language

The language edition of Wikipedia that the titles come from

Value

A list of arrays of page language objects, the same length as the passed list of titles. Each page language object has keys: code and name describe the language edition in which the linked page was found; while key and title identify the page on the linked wiki.

Details

If the function fails to find any interwiki links for the passed title, it will fail silently.

Examples

# Get language links for a single page on English Wikipedia
get_langlinks("Charles_Harpur")

# Get a tbl of pages from French Wikipedia using the Action API, then
# find their interwiki links
albums_des_beatles <- wiki_action_request(language = "fr") %>%
  query_generate_pages("categorymembers", gcmtitle="Catégorie:Album_des_Beatles", gcmtype="page") %>%
  query_page_properties("title") %>%
  retrieve_all() %>%
  dplyr::mutate(links = get_langlinks(title, language="fr"))

# links is a list column, so [tidyr] and [purrr] are useful
# Count the number of langlinks for each page
albums_des_beatles %>% dplyr::mutate(
  count = purrr::map_dbl(links, length)
)
#> # A tibble: 30 × 5
#>      pageid    ns title                                             links  count
#>       <int> <int> <chr>                                             <list> <dbl>
#>  1   314256     0 The Beatles (album)                               <list>    54
#>  2   466773     0 Abbey Road (album)                                <list>    58
#>  3   639219     0 A Hard Day's Night (album)                        <list>    53
#>  4   639231     0 Beatles for Sale                                  <list>    46
#>  5  3776338     0 Beatles '65                                       <list>    17
#>  6  4899948     0 The Beatles at the Hollywood Bowl                 <list>    14
#>  7  9342535     0 The Beatles' Long Tall Sally                      <list>     9
#>  8 11105806     0 Pochette de Sgt. Pepper's Lonely Hearts Club Band <list>     5
#>  9 13345678     0 The Beatles' First !                              <list>     8
#> 10 14065386     0 Les Beatles (album français)                      <list>     0
#> # … with 20 more rows
# Unnest the data frame to get one row per interwiki link, and extract
# the language code and page title for each interwiki link
albums_des_beatles %>%
  tidyr::unnest(links) %>%
  tidyr::hoist(links, "code", linked_title = "title")
#> # A tibble: 897 × 6
#>    pageid    ns title               code     linked_title          links       
#>     <int> <int> <chr>               <chr>    <chr>                 <list>      
#>  1 314256     0 The Beatles (album) af       The Beatles (album)   <named list>
#>  2 314256     0 The Beatles (album) ar       البيتلز (ألبوم)       <named list>
#>  3 314256     0 The Beatles (album) az       The Beatles (albom)   <named list>
#>  4 314256     0 The Beatles (album) be       The Beatles (альбом)  <named list>
#>  5 314256     0 The Beatles (album) be-x-old The Beatles (альбом)  <named list>
#>  6 314256     0 The Beatles (album) bg       The Beatles (албум)   <named list>
#>  7 314256     0 The Beatles (album) br       The Beatles (pladenn) <named list>
#>  8 314256     0 The Beatles (album) bs       The Beatles (album)   <named list>
#>  9 314256     0 The Beatles (album) ca       The Beatles (àlbum)   <named list>
#> 10 314256     0 The Beatles (album) ceb      The Beatles (Album)   <named list>
#> # … with 887 more rows