Build a query to enumerate the members of a category in Wikipedia

This function provides access to the CategoryMembers endpoint of the Action API. Categories in Wikipedia contain two main kinds of 'members': subcategories and pages.

Usage

query_category_members(
  category,
  namespace = NULL,
  type = c("file", "page", "subcat"),
  limit = 10,
  sort = c("sortkey", "timestamp"),
  dir = c("ascending", "descending", "newer", "older"),
  start = NULL,
  end = NULL,
  language = "en"
)

Arguments

category

The category to start from. Can be named by title or pageid.

namespace

Only return category members from the provided namespace

type

Alternative to namespace: the type of category member to return

limit

The number to return each batch. Max 500.

sort

How to sort the returned category members. 'timestamp' sorts them by the date they were included in the category; 'sortkey' by the category member's unique hexadecimal code

dir

The direction in which to sort them

start

If sort == 'timestamp', only return category members from after this date. The argument is parsed by lubridate::as_date()

end

If sort == 'timestamp', only return category members included in the category from before this date. The argument is parsed by lubridate::as_date()

language

The language edition of Wikipedia to query

Value

A request object of type generator/query/action_api/httr2_request, which can be passed to next_batch() or retrieve_all(). You can specify which properties to retrieve for each page using query_page_properties().

Examples

# Get the first 10 pages in 'Category:Physics' on English Wikipedia
physics_members <- query_category_members("Physics") %>% next_batch()
physics_members
#> <query_tbl>
#> Request URL:
#> https://en.wikipedia.org/w/api.php?action=query&format=json&formatversion=2&generator=categorymembers&gcmtitle=Category%3APhysics&gcmtype=file%7Cpage%7Csubcat&gcmlimit=10&gcmsort=sortkey&gcmdir=ascending
#> ℹ There are more results on the server. Retrieve them with `next_batch()` or `retrieve_all()`
#> ✔ Data complete for all records
#> # A tibble: 10 × 3
#>      pageid    ns title                        
#>       <int> <int> <chr>                        
#>  1    22939     0 Physics                      
#>  2    23479     0 Physicalism                  
#>  3   570922     0 Action at a distance         
#>  4   694942    14 Category:Physicists          
#>  5  1653925   100 Portal:Physics               
#>  6  1854369     0 Hydraulic head               
#>  7  5625591    14 Category:Physics awards      
#>  8 36477012    14 Category:Concepts in physics 
#>  9 49740128    14 Category:Subfields of physics
#> 10 70983414    14 Category:Physics by country  

# Retrieve interwiki links for the first 10 pages in the 'Beatles Albums'
# category on Danish Wikipedia
# TODO: This currently doesn't work, because incomplete batches aren't properly handled
beatles <- query_category_members("Beatles-album", language = "da") %>%
  query_page_properties("langlinks") %>%
  next_batch()