Search for insertions, deletions or relocations of text between two versions of a Wikipedia page

Any two revisions of a Wikipedia page can be compared using the 'diff' tool. The tool compares the 'from' revision to the 'to' revision, looking for insertions, deletions or relocations of text. This operation can be performed in any order, across any span of revisions.

Usage

get_diff(from, to, language = "en")

Arguments

from

Vector of revision ids

to

Vector of revision ids

language

Vector of two-letter language codes (will be recyled if length==1)

Value

A list the same length as from and to, comprising wikidiff2 responses, represented as R lists

Examples

# Compare revision 847170467 to 851733941 on English Wikipedia
diffs <- get_diff(847170467, 851733941)

# Each wikidiff2 response contains three main keys:
names(diffs[[1]])
#> [1] "from" "to"   "diff"
# [1] "from" "to" "diff"

# To see the text that has been inserted, deleted, or moved, look in the
# `diff` key, which contains a numbered list of differences
diffs[[1]]$diff[[2]]$text
#> [1] "Though ''Neuranethes spodopterodes'' is invasive in the regions where it has emerged as a pest, it is not an exotic invader, but a [[Species translocation|translocated species]], having been imported inadvertently from its natural range in more northerly regions of the country. In its original range the moth is not of horticultural importance, being controlled by natural enemies that as yet have neither been identified nor imported along with the host plants. In contrast the ''Agapanthus'' borer is of considerable concern in the South West, and its voracity is so impressive that the species shows promise as a possible control for invasive ''Agapanthus praecox'' in countries like New Zealand.<ref>M.D. Picker and M. Krüger. Spread and Impacts of the Agapanthus Borer (Neuranethes spodopterodes (Hampson, 1908), comb. nov.), a Translocated Native Moth Species (Lepidoptera: Noctuidae). African Entomology 2013 21 (1), 172-176</ref>"

# In this case, the diff is of type `0`, which means an insertion
diffs[[1]]$diff[[2]]$type
#> [1] 0

# The function is vectorised, so you can compare multiple pairs of revisions
# in a single call
# See diffs for the last five revisions of the Main Page
revisions <- wiki_action_request() %>%
  query_page_properties(
    "revisions",
    titles = "Main_Page", rvlimit = 5, rvprop = "ids", rvdir = "older"
  ) %>%
  perform_query_once() %>%
  tidyr::hoist(revisions, "parentid", "revid")
diffs <- get_diff(from = revisions$parentid, to = revisions$revid)