Perform a query using the MediaWiki Action API

next_result() sends exactly one request to the server.

next_batch() requests results from the server until data is complete the latest batch of pages in the result.

retrieve_all() keeps requesting data until all the pages from the query have been returned.

Usage

next_result(x)

next_batch(x)

retrieve_all(x)

Arguments

x

The query. Either a wiki_action_request or a query_tbl.

Value

A query_tbl containing results of the query. If x is a query_tbl, then the function will return a new data with the new data appended to it. If x is a wiki_action_request, then the returned query_tbl will contain the necessary data to supply future calls to next_result(), next_batch() or retrieve_all().

Details

It is rare that a query can be fulfilled in a single request to the server. There are two ways a query can be incomplete. All queries return a list of pages as their result. The result may be incomplete because not all the data for each page has been returned. In this case the batch is incomplete. Or the data may be complete for all pages, but there are more pages available on the server. In this case the query can be continued. Thus the three functions for next_result(), next_batch() and retrieve_all().

Examples

# Try out a request using next_batch(), then retrieve the rest of the
# results.
preview <- wiki_action_request() %>%
  query_by_title("Albert Einstein") %>%
  query_page_properties("categories") %>%
  next_batch()
preview
#> <final/query_tbl>
#> ✔ All results downloaded from server
#> ✔ Data complete for all records
#> # A tibble: 1 × 4
#>   pageid    ns title           categories        
#>    <int> <int> <chr>           <list>            
#> 1    736     0 Albert Einstein <tibble [147 × 2]>
all_results <- retrieve_all(preview)
all_results
#> <final/query_tbl>
#> ✔ All results downloaded from server
#> ✔ Data complete for all records
#> # A tibble: 1 × 4
#>   pageid    ns title           categories        
#>    <int> <int> <chr>           <list>            
#> 1    736     0 Albert Einstein <tibble [147 × 2]>