Explore Wikipedia's category system

These functions provide access to the CategoryMembers endpoint of the Action API.

query_category_members() builds a generator query to return the members of a given category.

build_category_tree() finds all the pages and subcategories beneath the passed category, then recursively finds all the pages and subcategories beneath them, until it can find no more subcategories.

Usage

query_category_members(
  .req,
  category,
  namespace = NULL,
  type = c("file", "page", "subcat"),
  limit = 10,
  sort = c("sortkey", "timestamp"),
  dir = c("ascending", "descending", "newer", "older"),
  start = NULL,
  end = NULL,
  language = "en"
)

build_category_tree(category, language = "en")

Arguments

.req

A query request object

category

The category to start from. query_category_members() accepts either a numeric pageid or the page title. build_category_tree() accepts a vector of page titles.

namespace

Only return category members from the provided namespace

type

Alternative to namespace: the type of category member to return. Multiple types can be requested using a character vector. Defaults to all.

limit

The number to return each batch. Max 500.

sort

How to sort the returned category members. 'timestamp' sorts them by the date they were included in the category; 'sortkey' by the category member's unique hexadecimal code

dir

The direction in which to sort them

start

If sort == 'timestamp', only return category members from after this date. The argument is parsed by lubridate::as_date()

end

If sort == 'timestamp', only return category members included in the category from before this date. The argument is parsed by lubridate::as_date()

language

The language edition of Wikipedia to query

Value

query_category_members(): A request object of type generator/query/action_api/httr2_request, which can be passed to next_batch() or retrieve_all(). You can specify which properties to retrieve for each page using query_page_properties().

build_category_tree(): A list containing two dataframes. nodes lists all the subcategories and pages found underneath the passed categories. edges records the connections between them. The source column gives the pageid of the parent category, while the target column gives the pageid of any categories, pages or files contained within the source category. The timestamp records the moment when the target page or subcategory was included in the source category. The two dataframes in the list can be passed to igraph::graph_from_data_frame for network analysis.

Examples

# Get the first 10 pages in 'Category:Physics' on English Wikipedia
physics_members <- wiki_action_request() %>%
  query_category_members("Physics") %>% next_batch()
physics_members
#> <complete/query_tbl>
#> ℹ There are more results on the server. Retrieve them with `next_batch()` or `retrieve_all()`
#> ✔ Data complete for all records
#> # A tibble: 10 × 3
#>      pageid    ns title                        
#>       <int> <int> <chr>                        
#>  1    22939     0 Physics                      
#>  2    23479     0 Physicalism                  
#>  3   570922     0 Action at a distance         
#>  4   694942    14 Category:Physicists          
#>  5  1653925   100 Portal:Physics               
#>  6  1854369     0 Hydraulic head               
#>  7  5625591    14 Category:Physics awards      
#>  8 36477012    14 Category:Concepts in physics 
#>  9 49740128    14 Category:Subfields of physics
#> 10 70983414    14 Category:Physics by country  

# Retrieve interwiki links for the first 10 pages in the 'Beatles Albums'
# category on Danish Wikipedia
# TODO: This currently doesn't work, because incomplete batches aren't properly handled
beatles <- wiki_action_request(language = "da") %>%
  query_category_members("Beatles-album") %>%
  query_page_properties("langlinks") %>%
  next_batch()

# Build the tree of all trees in Wikipedia
tree <- build_category_tree("Category:Trees (graph theory)")
#> ⠙ Walking subcategories: 1 done (423/s) | 3ms
#> ⠹ Walking subcategories: 2 done (22/s) | 93ms
#> ⠸ Walking subcategories: 3 done (18/s) | 167ms
#> ⠼ Walking subcategories: 4 done (15/s) | 261ms
#> ⠴ Walking subcategories: 5 done (15/s) | 335ms
#> ⠦ Walking subcategories: 6 done (15/s) | 413ms
#> ⠧ Walking subcategories: 7 done (14/s) | 488ms
#> ⠇ Walking subcategories: 8 done (14/s) | 556ms
#> ⠏ Walking subcategories: 9 done (14/s) | 645ms
tree
#> $nodes
#> # A tibble: 298 × 4
#>      pageid    ns title                         type 
#>       <int> <int> <chr>                         <chr>
#>  1 17606898    14 Category:Trees_(graph_theory) root 
#>  2    48560     0 Tree (graph theory)           page 
#>  3 44659935     0 Agreement forest              page 
#>  4 13251077     0 Arborescence (graph theory)   page 
#>  5  1410431     0 Bethe lattice                 page 
#>  6 29106883     0 Block graph                   page 
#>  7 48861658     0 Blossom tree (graph theory)   page 
#>  8 16823137     0 Branch-decomposition          page 
#>  9 68638522     0 Buchholz hydra                page 
#> 10 27258886     0 Caterpillar tree              page 
#> # ℹ 288 more rows
#> 
#> $edges
#> # A tibble: 318 × 3
#>      source   target timestamp           
#>       <int>    <int> <chr>               
#>  1 17606898    48560 2008-05-25T15:51:35Z
#>  2 17606898 44659935 2014-12-12T11:24:55Z
#>  3 17606898 13251077 2010-09-09T15:01:59Z
#>  4 17606898  1410431 2008-05-25T15:53:18Z
#>  5 17606898 29106883 2010-10-08T00:51:39Z
#>  6 17606898 48861658 2015-12-21T22:02:29Z
#>  7 17606898 16823137 2008-06-01T07:24:50Z
#>  8 17606898 68638522 2021-09-03T22:35:48Z
#>  9 17606898 27258886 2010-05-07T00:17:15Z
#> 10 17606898  1266713 2008-06-01T07:17:42Z
#> # ℹ 308 more rows
#> 

# For network analysis and visualisation, you can pass the category tree
# to igraph
tree_graph <- igraph::graph_from_data_frame(tree$edges, vertices = tree$nodes)
tree_graph
#> IGRAPH d36d565 DN-B 298 318 -- 
#> + attr: name (v/c), ns (v/n), title (v/c), type (v/c), timestamp (e/c)
#> + edges from d36d565 (vertex names):
#>  [1] 17606898->48560    17606898->44659935 17606898->13251077 17606898->1410431 
#>  [5] 17606898->29106883 17606898->48861658 17606898->16823137 17606898->68638522
#>  [9] 17606898->27258886 17606898->1266713  17606898->4557120  17606898->267355  
#> [13] 17606898->40006935 17606898->19716804 17606898->68857410 17606898->31104438
#> [17] 17606898->3606300  17606898->35471883 17606898->36344169 17606898->7196522 
#> [21] 17606898->2145273  17606898->4587190  17606898->53993905 17606898->1731488 
#> [25] 17606898->1650168  17606898->62951851 17606898->22192834 17606898->27561614
#> [29] 17606898->5140949  17606898->2185021  17606898->19310129 17606898->53828785
#> + ... omitted several edges