Specify which pages to generate from the Action API

See API:Lists for available generators. Most list actions in the MediaWiki API can also be used as generators. If you use a list action as a generator, then you will receive the data in the pages attribute of the response body, rather than under the name of the generator. You can query_generate_pages with query_page_properties to choose what properties of each page to return.

Usage

query_generate_pages(.req, generator, ...)

Arguments

.req

A httr2_request, e.g. generated by wiki_action_request

generator

The type of generator to return

...

Additional parameters to the generator

Value

An HTTP response: an S3 list with class httr2_request

Details

To set additional parameters to a generator, prepend the parameter with "g". For instance, to set a limit of 10 to the number of pages returned by the categorymembers generator, set the parameter gcmlimit = 10.

Examples

# Search for articles about seagulls

resp <- wiki_action_request() %>%
  query_generate_pages("search", gsrsearch="seagull") %>%
  httr2::req_perform() %>%
  httr2::resp_body_json()

resp$query$pages
#> [[1]]
#> [[1]]$pageid
#> [1] 45232
#> 
#> [[1]]$ns
#> [1] 0
#> 
#> [[1]]$title
#> [1] "Jonathan Livingston Seagull"
#> 
#> [[1]]$index
#> [1] 4
#> 
#> 
#> [[2]]
#> [[2]]$pageid
#> [1] 210481
#> 
#> [[2]]$ns
#> [1] 0
#> 
#> [[2]]$title
#> [1] "Gull"
#> 
#> [[2]]$index
#> [1] 2
#> 
#> 
#> [[3]]
#> [[3]]$pageid
#> [1] 210492
#> 
#> [[3]]$ns
#> [1] 0
#> 
#> [[3]]$title
#> [1] "Seagull (disambiguation)"
#> 
#> [[3]]$index
#> [1] 5
#> 
#> 
#> [[4]]
#> [[4]]$pageid
#> [1] 474370
#> 
#> [[4]]$ns
#> [1] 0
#> 
#> [[4]]$title
#> [1] "The Seagull"
#> 
#> [[4]]$index
#> [1] 3
#> 
#> 
#> [[5]]
#> [[5]]$pageid
#> [1] 887152
#> 
#> [[5]]$ns
#> [1] 0
#> 
#> [[5]]$title
#> [1] "Barbara Hershey"
#> 
#> [[5]]$index
#> [1] 7
#> 
#> 
#> [[6]]
#> [[6]]$pageid
#> [1] 2305766
#> 
#> [[6]]$ns
#> [1] 0
#> 
#> [[6]]$title
#> [1] "British Seagull"
#> 
#> [[6]]$index
#> [1] 9
#> 
#> 
#> [[7]]
#> [[7]]$pageid
#> [1] 6386157
#> 
#> [[7]]$ns
#> [1] 0
#> 
#> [[7]]$title
#> [1] "Seagull management"
#> 
#> [[7]]$index
#> [1] 6
#> 
#> 
#> [[8]]
#> [[8]]$pageid
#> [1] 12811837
#> 
#> [[8]]$ns
#> [1] 0
#> 
#> [[8]]$title
#> [1] "Seagull Camera"
#> 
#> [[8]]$index
#> [1] 10
#> 
#> 
#> [[9]]
#> [[9]]$pageid
#> [1] 24071169
#> 
#> [[9]]$ns
#> [1] 0
#> 
#> [[9]]$title
#> [1] "Tianjin Seagull"
#> 
#> [[9]]$index
#> [1] 8
#> 
#> 
#> [[10]]
#> [[10]]$pageid
#> [1] 58723600
#> 
#> [[10]]$ns
#> [1] 0
#> 
#> [[10]]$title
#> [1] "Bobby Seagull"
#> 
#> [[10]]$index
#> [1] 1
#> 
#>