Choose properties to return for pages from the action API

See API:Properties for a list of available properties. Many have additional parameters to control their behavior, which can be passed to this function as named arguments.

Usage

query_page_properties(.req, properties, ...)

Arguments

.req

A httr2_request, e.g. generated by wiki_action_request

properties

The properties to request, as either a character vector or a string with properties separated by "|"

...

Additional parameters to pass, e.g. to modify what is returned by the property request

Value

An HTTP response: an S3 list with class httr2_request

Details

query_page_properties can be used in combination with query_generate_pages to choose what information to retrieve about the generated pages. If you already have a list of page ids or titles, you can build a query using query_page_properties to collect information about them.

Examples

# Search for articles about seagulls and return their links to other
# language wikis

resp <- wiki_action_request() %>%
  query_generate_pages("search", gsrsearch="seagull") %>%
  query_page_properties("langlinks") %>%
  httr2::req_perform() %>%
  httr2::resp_body_json()

resp$query$pages
#> [[1]]
#> [[1]]$pageid
#> [1] 45232
#> 
#> [[1]]$ns
#> [1] 0
#> 
#> [[1]]$title
#> [1] "Jonathan Livingston Seagull"
#> 
#> [[1]]$index
#> [1] 3
#> 
#> [[1]]$langlinks
#> [[1]]$langlinks[[1]]
#> [[1]]$langlinks[[1]]$lang
#> [1] "ast"
#> 
#> [[1]]$langlinks[[1]]$title
#> [1] "Juan Salvador Gaviota"
#> 
#> 
#> [[1]]$langlinks[[2]]
#> [[1]]$langlinks[[2]]$lang
#> [1] "bg"
#> 
#> [[1]]$langlinks[[2]]$title
#> [1] "Джонатан Ливингстън Чайката"
#> 
#> 
#> [[1]]$langlinks[[3]]
#> [[1]]$langlinks[[3]]$lang
#> [1] "cs"
#> 
#> [[1]]$langlinks[[3]]$title
#> [1] "Jonathan Livingston Racek"
#> 
#> 
#> [[1]]$langlinks[[4]]
#> [[1]]$langlinks[[4]]$lang
#> [1] "de"
#> 
#> [[1]]$langlinks[[4]]$title
#> [1] "Die Möwe Jonathan"
#> 
#> 
#> [[1]]$langlinks[[5]]
#> [[1]]$langlinks[[5]]$lang
#> [1] "el"
#> 
#> [[1]]$langlinks[[5]]$title
#> [1] "Ο γλάρος Ιωνάθαν Λίβινγκστον"
#> 
#> 
#> [[1]]$langlinks[[6]]
#> [[1]]$langlinks[[6]]$lang
#> [1] "es"
#> 
#> [[1]]$langlinks[[6]]$title
#> [1] "Juan Salvador Gaviota"
#> 
#> 
#> [[1]]$langlinks[[7]]
#> [[1]]$langlinks[[7]]$lang
#> [1] "eu"
#> 
#> [[1]]$langlinks[[7]]$title
#> [1] "Jonatan Kaioa"
#> 
#> 
#> [[1]]$langlinks[[8]]
#> [[1]]$langlinks[[8]]$lang
#> [1] "fa"
#> 
#> [[1]]$langlinks[[8]]$title
#> [1] "جاناتان مرغ دریایی"
#> 
#> 
#> [[1]]$langlinks[[9]]
#> [[1]]$langlinks[[9]]$lang
#> [1] "fi"
#> 
#> [[1]]$langlinks[[9]]$title
#> [1] "Lokki Joonatan"
#> 
#> 
#> [[1]]$langlinks[[10]]
#> [[1]]$langlinks[[10]]$lang
#> [1] "fr"
#> 
#> [[1]]$langlinks[[10]]$title
#> [1] "Jonathan Livingston le goéland"
#> 
#> 
#> 
#> 
#> [[2]]
#> [[2]]$pageid
#> [1] 210481
#> 
#> [[2]]$ns
#> [1] 0
#> 
#> [[2]]$title
#> [1] "Gull"
#> 
#> [[2]]$index
#> [1] 1
#> 
#> 
#> [[3]]
#> [[3]]$pageid
#> [1] 210492
#> 
#> [[3]]$ns
#> [1] 0
#> 
#> [[3]]$title
#> [1] "Seagull (disambiguation)"
#> 
#> [[3]]$index
#> [1] 6
#> 
#> 
#> [[4]]
#> [[4]]$pageid
#> [1] 474370
#> 
#> [[4]]$ns
#> [1] 0
#> 
#> [[4]]$title
#> [1] "The Seagull"
#> 
#> [[4]]$index
#> [1] 2
#> 
#> 
#> [[5]]
#> [[5]]$pageid
#> [1] 712580
#> 
#> [[5]]$ns
#> [1] 0
#> 
#> [[5]]$title
#> [1] "Seagull Monument"
#> 
#> [[5]]$index
#> [1] 10
#> 
#> 
#> [[6]]
#> [[6]]$pageid
#> [1] 887152
#> 
#> [[6]]$ns
#> [1] 0
#> 
#> [[6]]$title
#> [1] "Barbara Hershey"
#> 
#> [[6]]$index
#> [1] 7
#> 
#> 
#> [[7]]
#> [[7]]$pageid
#> [1] 6386157
#> 
#> [[7]]$ns
#> [1] 0
#> 
#> [[7]]$title
#> [1] "Seagull management"
#> 
#> [[7]]$index
#> [1] 5
#> 
#> 
#> [[8]]
#> [[8]]$pageid
#> [1] 12811837
#> 
#> [[8]]$ns
#> [1] 0
#> 
#> [[8]]$title
#> [1] "Seagull Camera"
#> 
#> [[8]]$index
#> [1] 9
#> 
#> 
#> [[9]]
#> [[9]]$pageid
#> [1] 47120951
#> 
#> [[9]]$ns
#> [1] 0
#> 
#> [[9]]$title
#> [1] "The Seagull (2018 film)"
#> 
#> [[9]]$index
#> [1] 8
#> 
#> 
#> [[10]]
#> [[10]]$pageid
#> [1] 58723600
#> 
#> [[10]]$ns
#> [1] 0
#> 
#> [[10]]$title
#> [1] "Bobby Seagull"
#> 
#> [[10]]$index
#> [1] 4
#> 
#>